requirements.txt
tensorflow>=2.15.0
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
flask>=3.0.0
flask-cors>=4.0.0
pillow>=10.0.0
gunicorn>=21.2.0
fpdf2>=2.7.0
